Add vega
- Move all packages from overlays/ and pkgs/ to vega/pkgs/. They are now
  accessed with `vega.pkgs` instead of directly through `pkgs`.
- Move custom lib from lib/ to vega/lib/. It is now accessed with
  `vega.lib` instead of `muhlib`. Also, all available functions have
  been merged into a single file.
- Add nixpkgs config to its own file (modules/nixpkgs.nix).

Vega is a whole flake for my custom lib, custom packages and some flake
packages. I don't know yet if it's a good idea for vega to be a whole
flake instead of a simple namespace. Also "vega" is a weird name, I'm
considering "my", "forklift" or "blueprint" instead with diminutive for
the namespace.